Вынесено сохранение рандомной подборки из алгоритма рандомной подборки. Таким образом алгоритм подбирает (и только), а менеджер смотрит есть ли подборка и нужно ли подбирать заново
