+++ Services/Managers/FilmManager.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Core.Enums;
 
 namespace Services.Managers
 {
@@ -15,18 +16,21 @@ namespace Services.Managers
         private IRepository<Film> _filmsRepo;
         private IRepository<Account> _usersRepo;
         private IRepository<UserFilm> _userFilmsRepo;
+        private readonly IRepository<SelectionList> _selectionListRepo;
         private ISameUsersAlgorithm _specifityFilmSelector;
         private readonly IRandomFilmsAlgorithm _randomFilmsAlgorithm;
 
         public FilmManager(IRepository<Film> films,
                             IRepository<Account> users, 
                             IRepository<UserFilm> userFilms,
+                            IRepository<SelectionList> selectionListRepo,
                             ISameUsersAlgorithm specifityFilmSelector,
                             IRandomFilmsAlgorithm randomFilmsAlgorithm)
         {
             _filmsRepo = films;
             _usersRepo = users;
             _userFilmsRepo = userFilms;
+            _selectionListRepo = selectionListRepo;
             _specifityFilmSelector = specifityFilmSelector;
             _randomFilmsAlgorithm = randomFilmsAlgorithm;
         }
@@ -53,7 +57,31 @@ namespace Services.Managers
 
         public async Task<IList<Film>> GetRandomShakedFilms(string userId = null)
         {
-            return await _randomFilmsAlgorithm.GetFilms(userId);
+            //Проверяем существует ли подборка и актуальна ли она (например не больше чем месячной давности)
+            var list = _selectionListRepo
+                .Get()
+                .Include(sl => sl.FilmSelectionLists)
+                    .ThenInclude(fsl => fsl.Film)
+                        .ThenInclude(f => f.Preview)
+                .FirstOrDefault(fl => fl.UserId == userId
+                    && fl.AlgorithmType == AlgorithmType.RandomAlgorithm);
+
+            //Подборка существует и она нынешнего месяца
+            if (list != null && DateTime.UtcNow.Month == list.CreatedOn.Month)
+                return list.FilmSelectionLists.Select(fsl => fsl.Film).ToList();
+
+            //Подборка существует но она просрочена
+            else if (list != null)
+            {
+                _selectionListRepo.Delete(list.Id);
+                await _selectionListRepo.SaveAsync();
+            }
+
+            // Не существует никакой
+            var films = await _randomFilmsAlgorithm.GetFilms(userId);
+            if (!string.IsNullOrEmpty(userId))
+                await CreateList(userId, films);
+            return films;
         }
 
         public IList<Genre> GetGenres(Guid id)
@@ -114,6 +142,25 @@ namespace Services.Managers
             var result = _filmsRepo.Delete(id);
             await _filmsRepo.SaveAsync();
             return result;
+        }
+
+        private async Task CreateList(string userId, IList<Film> films)
+        {
+            var filmSelectionLists = films.Select((film, index) => new FilmSelectionList
+            {
+                FilmId = film.Id,
+                Order = index
+            }).ToList();
+
+            var selectionList = new SelectionList
+            {
+                UserId = userId,
+                AlgorithmType = AlgorithmType.RandomAlgorithm,
+                CreatedOn = DateTime.UtcNow,
+                FilmSelectionLists = filmSelectionLists
+            };
+            await _selectionListRepo.CreateAsync(selectionList);
+            await _selectionListRepo.SaveAsync();
         }
     }
 }
